+++ lib/cell.rb
@@ -22,7 +22,7 @@ end
 
 require 'cell/caching'
 require 'cell/caching/notification'
-require 'cell/builder'
+require 'uber/builder'
 require 'cell/prefixes'
 require 'cell/self_contained'
 require 'cell/layout'